requirements.txt
asgiref==3.7.2
Django==5.0.1
django-modeltranslation==0.18.11
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
pillow==10.2.0
sqlparse==0.4.4
typing-extensions==4.9.0
tzdata==2023.4
